requirements.txt
ccxt
pandas
numpy
TA-Lib
backtrader
websocket-client
pyyaml
python-dotenv